Support definitions pattern for when $ref's are't pre-resolved
Sometimes it isn't ideal to have to resolve refs before running
openapi-schema-to-json-schema. In my case I need to run type generators
on individual schema files, not the combined schemas I'd get from
running something like json-ref-resolver.

This adds a new option called definitionKeywords that takes an array of
keywords to treat as "definition objects". Each property in the
"definition object" will be treated as a subschema and have all
conversions run on them.